add missing FollowButton
